feat(restaurantList): 음식점 리스트 도메인 추가 Co-authored-by: skiende74 <skiende74@gmail.com>
